Streamline distributed locking.
We now lock at the level of the object being linked. This makes it easy to horizontally scale linking as required by allowing disjoint clusters to be processed independently.
The key observation is that main decision is merge or create. You don't want to merge into a cluster that is potentially being created or updated so if you block to a locked element
we simply skip processing that element. Still need to handle the merge case where two candidates that don't block to each other, but have a cluster in common, may attempt to merge into
the same cluster.