Refine reading and writing of data.
- Simplified query for reading data.
- Auto-commit is enabled for writing with proper transactional mechanics.
- Finalizing a write transaction propagates writes to linked entities. This code will also be useful in the future when needing to propagate changes from items that are linked together.

Due to changes in aggregation we know longer need to accumulate properties in datastore as this handled by Postgres. I left in the group by origin_id for backwards compatibility, but this is a short term fix
as it is not ideal to have a read path just for linking indexing which doesn't reflect data this is loaded. Long term, I think linking indexing should be able to use aggregation over detail objects in order to compute individual components for indexing. Not sure why this is necessary.