Implement multiple db support for graph.
For simplicity we write an edges to the resolved dbs for each src, dst, edge entity set triplet.

The risk here is that if service terminates in the middle of a write, data may not be committed across all different databases.

A future optimization maybe capturing all edges on primary database first along with a transaction flag, such that we can propagate any missing edges by filtering on flag.