Fixes the issue with creation of associations.
Associations created via the API were not showing up as they are frequently created as empty entities. Normally, an entity always has it's id property created, but after we added multiple database support that path
was missed as it was in entity key id generation service. An entity with no properties in the data table should up as a null neighbor, which caused it to be ignored by neighbors queries.

This fixes the root issue, so that association entities will always have their id entry created inside their respective data tables.